requirements.txt
torch
torchvision
mmsegmentation
mmengine
openmim
opencv-python
scikit-learn
pillow
numpy
tqdm
pyyaml
requests
